* upload sistemi değiştirildi, artık dosyalar private ve public olarak ayrılabiliyor. Private dosyalar fileeditor'de görünmüyor. * upload işlemi, dizin ve dosya işlemleri yapan fonksiyonlardan gerekli görünenlere access_type parametresi eklendi. access_type public veya private değerlerini alabilir. bu değerlere göre ilgili dosya veya dizinler private oluyorlar. * ufak hatalar giderilip güncellemeler yapıldı.

+++ system/classes/DIRECTORY.php
@@ -2,30 +2,43 @@
 class PA_DIRECTORY extends PA_FILE
 {
 	private $table;
+    private $public_root;
+    private $private_root;
 	
-	function PA_DIRECTORY()
-	{
+	function PA_DIRECTORY(){
 		parent::PA_FILE();
 		
 		$this->table = $this->tables->directory;
 		$this->table_file = $this->tables->file;
+
+        global $public_uploadurl;
+        global $private_uploadurl;
+
+        $this->public_root = $public_uploadurl;
+        $this->private_root = $private_uploadurl;
 	}
-	
-	function createDirectory($name, $parent_id=-1){
-		global $uploadurl;
+
+    /**
+     * @param $name dizin adı
+     * @param $parent_id parent dizinin id'si
+     * @param string $access_type dizine erişim tipini belirtir, "public" veya "private" olabilir.
+     * @return bool|string
+     */
+    function createDirectory($name, $parent_id=-1, $access_type="public"){
+        $root = $this->{$access_type . "_root"};
 		
-		if($dir = $this->selectDirectoryByNameAndParent($parent_id, $name)){
+		if($dir = $this->selectDirectoryByNameAndParent($parent_id, $name, $access_type)){
             return $dir->directory_id;
         }
         else{
-			// database e kaydedilecek şekilde dizin bilgisini hesapla
+			// database'e kaydedilecek şekilde dizin bilgisini hesapla
 			if($parent_directory = $this->selectDirectoryById($parent_id)) // Üst dizin varsa 
-				$directory = preg_replace("/^" . preg_quote($uploadurl, "/") . "/", "", $parent_directory->directory) . $name . "/";
+				$directory = preg_replace("/^" . preg_quote($root, "/") . "/", "", $parent_directory->directory) . $name . "/";
 			else
 				$directory = $name . "/";
 			
-			// dosyayı oluşturmada kullanmak için tam dizin bilgisini hespla
-			$full_path = $uploadurl . $directory;
+			// dosyayı oluşturmada kullanmak için tam dizin bilgisini hesapla
+			$full_path = $root . $directory;
 			
 			// dizinleri oluştur ve yetkilerini ata
 			if(!file_exists($full_path)){
@@ -38,13 +51,11 @@ class PA_DIRECTORY extends PA_FILE
 			}
 			
 			// database'e kaydet ve sonucu döndür
-			return $this->insert($this->table,array("parent_id"=>$parent_id,"name"=>$name,"directory"=>$directory));
+			return $this->insert($this->table, compact("parent_id", "name", "directory", "access_type"));
 		}
 	}	
 	
 	function updateDirectory($directory_id, $new_name){
-		global $uploadurl;
-
         $directory = $this->selectDirectoryById($directory_id);
         $selected_directory = $this->selectDirectoryByNameAndParent($directory->parent_id, $new_name);
 
@@ -62,16 +73,18 @@ class PA_DIRECTORY extends PA_FILE
 			$error = false;
 			
 			// Önce database de olmayan ve olmaması gereken ama bizim kolaylık olsun diye database den seçim esnasında CONCAT() ile eklediğimiz
-			// uploadurl'i temizle. Çünkü database de uploadurl bu şekilde ekli olarak kayıtlı değil ve aramayı doğru yapabilmek için uploadurl'i
+			// root url'i temizle. Çünkü database de root url bu şekilde ekli olarak kayıtlı değil ve aramayı doğru yapabilmek için root url'i
 			// temizlememiz gerekiyor.
-			$search_directory = preg_replace(("/^" . preg_quote($uploadurl, "/") . "/"), "", $directory->directory);
+
+            $root = $this->{$directory->access_type . "_root"};
+
+			$search_directory = preg_replace(("/^" . preg_quote($root, "/") . "/"), "", $directory->directory);
 			
-			// Şimdi uploadurl değeri temizlenmiş url içinden directory adını yenisiyle değiştir.
+			// Şimdi root url'i temizlenmiş url içinden directory adını yenisiyle değiştir.
 			$replace_directory = preg_replace(("/" . preg_quote($directory->name, "/") . "\/?$/"), $new_name . "/", $search_directory);
 
 			// varsa güncellenecek dizinin altında bulunan dosyaların url'lerini güncelle
-			if($files = $this->get_rows("SELECT file_id, url FROM {$this->table_file} WHERE directory_id > 0 AND url LIKE ? '%'",array($search_directory)))
-			{
+			if($files = $this->get_rows("SELECT file_id, url FROM {$this->table_file} WHERE directory_id > 0 AND access_type=? AND url LIKE ? '%'",array($directory->access_type, $search_directory))){
 				$file_count = sizeof($files);
 					
 				for($i=0; $i<$file_count; $i++){
@@ -82,8 +95,7 @@ class PA_DIRECTORY extends PA_FILE
 			}
 			
 			// varsa güncellenecek dizinin altındaki tüm dizinlerin directory değerini güncelle
-			if($directories = $this->get_rows("SELECT directory_id, directory FROM {$this->table} WHERE parent_id > 0 AND directory LIKE ? '%'", array($search_directory)))
-			{
+			if($directories = $this->get_rows("SELECT directory_id, directory FROM {$this->table} WHERE parent_id > 0 AND access_type=? AND directory LIKE ? '%'", array($directory->access_type, $search_directory))){
 				$directory_count = sizeof($directories);
 				
 				for($i=0; $i<$directory_count; $i++){
@@ -95,7 +107,7 @@ class PA_DIRECTORY extends PA_FILE
 			
 			// Son olarak hata yok ise istenen dizini güncelle
 			if(!$error){
-				if(rename($directory->directory, $uploadurl . $replace_directory)){
+				if(rename($directory->directory, $root . $replace_directory)){
 					return $this->execute("UPDATE {$this->table} SET name=?, directory=? WHERE directory_id=?", array($new_name, $replace_directory, $directory_id));
 				}
 				else{
@@ -107,23 +119,31 @@ class PA_DIRECTORY extends PA_FILE
 		}
 	}
 	
-	function selectDirectoryById($directory_id)
-	{
-		global $uploadurl;
-		return $this->get_row("SELECT *, CONCAT('{$uploadurl}', directory) AS directory FROM {$this->table} WHERE directory_id=?",array($directory_id));
+	function selectDirectoryById($directory_id){
+        if($directory = $this->get_row("SELECT * FROM {$this->table} WHERE directory_id=?", array($directory_id))){
+            $root = $this->{$directory->access_type . "_root"};
+            $directory->directory = $root . $directory->directory;
+
+            return $directory;
+        }
+        else{
+            return false;
+        }
+
 	}
 	
-	function selectDirectoryByNameAndParent($parent_id, $name)
-	{
-		global $uploadurl;
-		return $this->get_row("SELECT *, CONCAT('{$uploadurl}', directory) AS directory FROM {$this->table} WHERE parent_id=? AND name=?",array($parent_id, $name));
+	function selectDirectoryByNameAndParent($parent_id, $name, $access_type = "public"){
+        $root = $this->{$access_type . "_root"};
+
+        // directory_id hem public hem private dizinlerde -1 olabilir ve her dizinde de aynı isimde dosya bulunabilir. karısıkligi onlemek icin access_type degerini kullaniyoruz
+		return $this->get_row("SELECT *, CONCAT('{$root}', directory) AS directory FROM {$this->table} WHERE parent_id=? AND name=? AND access_type=?",array($parent_id, $name, $access_type));
 	}
 
     // parametre olarak verilen dizini parent dizinleri ile birlikte oluşturur.
-    function createDirectoryByPath($directory_path){
-        global $uploadurl;
+    function createDirectoryByPath($directory_path, $access_type="public"){
+        $root = $this->{$access_type . "_root"};
 
-        $fixed_path = preg_replace("/^.*?" . preg_quote($uploadurl, "/") . "/", "", $directory_path);
+        $fixed_path = preg_replace("/^.*?" . preg_quote($root, "/") . "/", "", $directory_path);
         $parent_id = -1;
 
         $path_array = explode("/", $fixed_path);
@@ -133,7 +153,7 @@ class PA_DIRECTORY extends PA_FILE
             $directory_name = $path_array[$i];
 
             if(strlen($directory_name) > 0){
-                if(!$parent_id = $this->createDirectory($directory_name, $parent_id)){
+                if(!$parent_id = $this->createDirectory($directory_name, $parent_id, $access_type)){
                     return false;
                 }
             }
@@ -142,46 +162,46 @@ class PA_DIRECTORY extends PA_FILE
         return $parent_id;
     }
 	
-	function listDirectoriesByParentId($parent_id)
-	{
-		global $uploadurl;
-		return $this->get_rows("SELECT *, CONCAT('$uploadurl', directory) AS directory FROM {$this->table} WHERE parent_id=?",array($parent_id));
+	function listDirectoriesByParentId($parent_id, $access_type = "public"){
+        $root = $this->{$access_type . "_root"};
+
+		return $this->get_rows("SELECT *, CONCAT('$root', directory) AS directory FROM {$this->table} WHERE parent_id=? AND access_type=?", array($parent_id, $access_type));
 	}
 	
-	function listFavouritedDirectories()
-	{
-		global $uploadurl;
-		return $this->get_rows("SELECT *, CONCAT('$uploadurl', directory) AS directory FROM {$this->table} WHERE is_favourite>0",null);
+	function listFavouritedDirectories($access_type = "public"){
+        $root = $this->{$access_type . "_root"};
+
+		return $this->get_rows("SELECT *, CONCAT('$root', directory) AS directory FROM {$this->table} WHERE is_favourite>0 AND access_type=?", array($access_type));
 	}
 	
-	function setDirectoryFavouriteStatus($directory_id, $status = 1)
-	{
+	function setDirectoryFavouriteStatus($directory_id, $status = 1){
 		return $this->execute("UPDATE {$this->table} SET is_favourite=? WHERE directory_id=?", array($status, $directory_id));
 	}
 	
-	function generateFileTreeHtmlByParentId($parent_id)
-	{
-		$dirs = $this->get_rows("SELECT * FROM {$this->table} WHERE parent_id=?", array($parent_id));
-		
-		$dirHtml = '<ul class="fileTree">';
-		foreach($dirs as $d)
-		{
-			$dirHtml .= '<li class="%s" directory_id="' . $d->directory_id . '">';
-			$dirHtml .= '<icon></icon><span class="name">' . $d->name . '</span>';
-			if($subdirs = $this->get_rows("SELECT * FROM {$this->table} WHERE parent_id=?",array($d->directory_id)))
-			{
-				$empty = "";
-				$dirHtml .= $this->generateFileTreeHtmlByParentId($d->directory_id);
-			}
-			else
-				$empty = "empty";
-			$dirHtml .= '</li>';
-			
-			$dirHtml = sprintf($dirHtml,$empty);
-			
-		}
-		$dirHtml .= '</ul>';
-		
+	function generateFileTreeHtmlByParentId($parent_id, $access_type="public"){
+		if($dirs = $this->get_rows("SELECT * FROM {$this->table} WHERE parent_id=? AND access_type=?", array($parent_id, $access_type))){
+            $dirHtml = '<ul class="fileTree">';
+            foreach($dirs as $d){
+
+                if($subTree = $this->generateFileTreeHtmlByParentId($d->directory_id)){
+                    $className = "";
+                }
+                else{
+                    $className = "empty";
+                }
+
+                $dirHtml .= '<li class="' . $className . '" directory_id="' . $d->directory_id . '">';
+                $dirHtml .= '<icon></icon><span class="name">' . $d->name . '</span>';
+                $dirHtml .= $subTree;
+                $dirHtml .= '</li>';
+            }
+            $dirHtml .= '</ul>';
+        }
+        else{
+            $dirHtml = "";
+        }
+
+
 		return $dirHtml;
 	}
 	
@@ -191,15 +211,13 @@ class PA_DIRECTORY extends PA_FILE
 	 * onların alt dosyalarının silinebilmesi için kullanılır.
 	 * 
 	 * */
-	function deleteDirectory($directory_id)
-	{
-		global $uploadurl;
-		
+	function deleteDirectory($directory_id){
 		$d = $this->selectDirectoryById($directory_id);
+        $root = $this->{$d->access_type . "_root"};
 		
 		if($this->deleteDirectoryCompletely($d->directory)) // dizin içindeki tüm dosya ve klasörleri sil
 		{
-			$d->directory = preg_replace(("/^" . preg_quote($uploadurl, "/") . "/"), "", $d->directory);
+			$d->directory = preg_replace(("/^" . preg_quote($root, "/") . "/"), "", $d->directory);
 			
 			// Dosyaların thumbnaillerini sil
 			if($files = $this->get_rows("SELECT * FROM {$this->table_file} WHERE url LIKE ? '%'",array($d->directory))){
@@ -211,9 +229,10 @@ class PA_DIRECTORY extends PA_FILE
 			}
 			
 			// Dosya ve dizin bilgilerini database den sil
-			if(!$this->execute("DELETE FROM {$this->table_file} WHERE access_type != 'system' AND url LIKE ? '%'",array($d->directory)) ||
-				!$this->execute("DELETE FROM {$this->table} WHERE directory LIKE ? '%'",array($d->directory)))
-			{
+			if(!$this->execute("DELETE FROM {$this->table_file} WHERE access_type=? AND url LIKE ? '%'",array($d->access_type, $d->directory))){
+                return false;
+            }
+            else if(!$this->execute("DELETE FROM {$this->table} WHERE access_type=? AND directory LIKE ? '%'",array($d->access_type, $d->directory))){
 				return false;
 			}
 		}
@@ -226,10 +245,8 @@ class PA_DIRECTORY extends PA_FILE
 	 * girilen dizini içerikleri ile birlikte siler, herhangi bir database işlemi yapmaz, işlemi sadece dosya bazında gerçekleştirir
 	 * @param string $directory
 	 */
-	private function deleteDirectoryCompletely($directory)
-	{
-		if(!file_exists($directory))
-		{
+	private function deleteDirectoryCompletely($directory){
+		if(!file_exists($directory)){
 			return true;
 		}
 		else if($contents = scandir($directory))
@@ -251,6 +268,9 @@ class PA_DIRECTORY extends PA_FILE
 			
 			return rmdir($directory);
 		}
+        else{
+            return true;
+        }
 	}
 
     /**
@@ -258,30 +278,27 @@ class PA_DIRECTORY extends PA_FILE
      * veritabanına ekler, kaydı olupta kendisi olmayan dizinlerin bilgilerini veritabanından siler
      * @return bool
      */
-    function synchronizeDirectories(){
-        global $ADMIN;
-        global $uploadurl;
-
-        $table_directory =  $ADMIN->DB->tables->directory;
+    function synchronizeDirectories($access_type = "public"){
+        $root = $this->{$access_type . "_root"};
 
         // dizinleri database'de ara ve olmayanları database'e ekle.
-        $directories = $this->listWholeDirectoriesInPath($uploadurl);
+        $directories = $this->listWholeDirectoriesInPath($root);
         $directory_amount = sizeof($directories);
         $error = false;
 
         for($i=0; $i<$directory_amount; $i++){
-            if(!$this->createDirectoryByPath($directories[$i])){
+            if(!$this->createDirectoryByPath($directories[$i], $access_type)){
                 $error = true;
             }
         }
 
         // database'deki dizinleri dosya sisteminde ara ve olmayanları database'den sil
-        $directories = $ADMIN->DB->get_rows("SELECT * FROM {$table_directory}");
+        $directories = $this->get_rows("SELECT * FROM {$this->table} WHERE access_type=?", array($access_type));
         foreach($directories as $d){
-            if(!is_dir($uploadurl . $d->directory)){
+            if(!is_dir($root . $d->directory)){
                 // eğer dizin yoksa veritabanından sil, dosya silme işlemi riskli olacağı için
                 // sistemin DIRECTORY class'ındaki fonksiyonu kullanma burda bir query çalıştır.
-                $ADMIN->DB->execute("DELETE FROM {$table_directory} WHERE directory_id=?", array($d->directory_id));
+                $this->execute("DELETE FROM {$this->table} WHERE directory_id=?", array($d->directory_id));
             }
         }
 
@@ -289,7 +306,7 @@ class PA_DIRECTORY extends PA_FILE
     }
 
     /**
-     * upload/files/ dizini içindeki herhangi bir dizinin içinde bulunan tüm dizinleri yine tüm alt dizinleri ile birlikte döndürür
+     * bir dizin içindeki herhangi bir dizinin içinde bulunan tüm dizinleri yine tüm alt dizinleri ile birlikte döndürür
      * @param $directory_path
      * @return array
      */
@@ -321,10 +338,10 @@ class PA_DIRECTORY extends PA_FILE
      * @param $directory_path
      * @return mixed
      */
-    function checkIfDirectoryExistInDbByPath($directory_path){
-        global $uploadurl;
+    function checkIfDirectoryExistInDbByPath($directory_path, $access_type="public"){
+        $root = $this->{$access_type . "_root"};
 
-        $fixed_path = preg_replace("/^" . preg_quote($uploadurl, "/") . "/", "", $directory_path);
+        $fixed_path = preg_replace("/^" . preg_quote($root, "/") . "/", "", $directory_path);
 
         return $this->get_row("SELECT * FROM {$this->table} WHERE directory=?", array($fixed_path));
     }
